Remove manual build of rcapsdump in spec file.
I was confused thinking that cmake was building rcapsdump, when it was
actually the spec file. The spec file had a manual list of libraries
that was incomplete. Now the build is setup using cmake configuration
so the correct libraries get included and the manual build was removed
from the spec file.